pyproject.toml
[project]
name = "manosaba_text_box"
version = "1.2.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyclip>=0.7.0",
  "pynput>=1.7.6",
  "rich",
  "textual",
  "pyyaml",
  "Pillow>=12.0.0",
  "pyperclip>=1.11.0",
  "psutil>=5.9.0",
  "pilmoji>=2.0.4",
  "emoji>=1.7.0",
  "pywin32>=311; platform_system == \"Windows\"",
]

[tool.uv]
